requirements.txt
faster-whisper>=0.9.0
pyaudio>=0.2.11
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
PySide6>=6.6.0 